refactor(Refactor hydration handling and state management.): Refactor hydration handling using useRef to prevent unnecessary re-renders and avoid hydration errors.
The component previously relied on useState to manage hydration state, which led to hydration errors when the client and server render states were mismatched. In this update, we’ve removed the use of useState for tracking the mounted state and replaced it with useRef. The useRef ensures that the component correctly identifies when it’s mounted and prevents unnecessary re-renders, avoiding hydration issues. This change improves the performance and reliability of the component during the client-side rendering phase, ensuring a smooth user experience.

BREAKING CHANGE: No breaking changes. The component continues to function with the same logic, but the handling of hydration errors has been optimized.

Closes Closes #9 https://github.com/PatternedAI/demo-iframe/issues/9